feat: replace old navigation-menu ui component with most recent version
